Refactor reactor; Begin highlighter integration
