+++ packages/reactor/index.ts
@@ -2,5 +2,8 @@ export type {
 	Modification,
 	ModificationRequest,
 	AppliedModifications,
+	DomJsonExportNode,
+	TimeStampReference,
+	Highlighter,
 } from "./interfaces";
 export { Reactor } from "./reactor";
